fix: compile managed conn on mobile platforms
